+++ src/test/java/seedu/address/integration/SelectCommandTest.java
@@ -1,0 +1,42 @@
+package seedu.address.integration;
+
+import static org.junit.Assert.assertEquals;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import java.util.List;
+
+import org.junit.Test;
+
+import seedu.address.logic.commands.SelectCommand;
+import seedu.address.model.AddressBook;
+import seedu.address.model.person.Person;
+
+public class SelectCommandTest extends AddressBookIntegrationTest {
+
+    @Test
+    public void execute_selectInvalidArgsFormat_errorMessageShown() throws Exception {
+        String expectedMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT, SelectCommand.MESSAGE_USAGE);
+        assertIncorrectIndexFormatBehaviorForCommand("select", expectedMessage);
+    }
+
+    @Test
+    public void execute_selectIndexNotFound_errorMessageShown() throws Exception {
+        assertIndexNotFoundBehaviorForCommand("select");
+    }
+
+    @Test
+    public void execute_select_jumpsToCorrectPerson() throws Exception {
+        TestDataHelper helper = new TestDataHelper();
+        List<Person> threePersons = helper.generatePersonList(3);
+
+        AddressBook expectedAB = helper.generateAddressBook(threePersons);
+        helper.addToModel(model, threePersons);
+
+        assertCommandSuccess("select 2",
+            String.format(SelectCommand.MESSAGE_SELECT_PERSON_SUCCESS, 2),
+            expectedAB,
+            expectedAB.getPersonList());
+        assertEquals(1, targetedJumpIndex);
+        assertEquals(model.getFilteredPersonList().get(1), threePersons.get(1));
+    }
+}
